back_end/requirements.txt
Flask==2.2.3
Flask_Cors==3.0.10
Flask_Mail==0.9.1
flask_sqlalchemy==3.0.3
PyJWT==2.6.0
pymongo==4.3.3
redis==4.5.4
SQLAlchemy==2.0.6
